requirements.txt
validators==0.18.2
discord.py==1.6.0
youtube_dl==2021.4.7
discord==1.7.3
python-dotenv==0.19.0
